Move and rename ToPendingTransition
